move SBML utility functions to a new file
